third_party/WebKit/LayoutTests/media/controls-timeline-with-controller.html

<!doctype html>
<html>
    <head>
        <title>media controls timeline with MediaController</title>
        <script src="media-file.js"></script>
        <script src="media-controls.js"></script>
        <script src="video-test.js"></script>
    </head>
    <body>
        <video controls></video>
        <script>
            findMediaElement();
            video.controller = new MediaController();
            video.src = findMediaFile("video", "content/test");
            waitForEvent("loadedmetadata", function()
            {
                testExpected("video.controller.currentTime", 0);
                testExpected("video.currentTime", 0);
                // Wait for the seek initiated by the "bringing up to speed"
                // step to complete.
                waitForEventOnce('seeked', seeked);
            });

            var seeked = function() {
                testExpected("video.seeking", false);

                // click the middle of the timeline
                var coords = mediaControlsButtonCoordinates(video, "timeline");
                eventSender.mouseMoveTo(coords[0], coords[1]);
                eventSender.mouseDown();
                eventSender.mouseUp();

                testExpected("video.controller.currentTime / video.controller.duration", 0.4, '>');
                testExpected("video.controller.currentTime / video.controller.duration", 0.6, '<');
                testExpected("video.currentTime / video.duration", 0.4, '>');
                testExpected("video.currentTime / video.duration", 0.6, '<');
                testExpected("video.seeking", true);

                endTest();
            }
        </script>
    </body>
</html>
